# Conflicts:
#	goods_receiving/src/main/java/com/lenta/bp9/di/AppComponent.kt
#	goods_receiving/src/main/java/com/lenta/bp9/platform/navigation/ScreenNavigator.kt
#	goods_receiving/src/main/res/values/strings.xml

+++ Shared/src/main/java/com/lenta/shared/utilities/databinding/TextViewDataBindingAdapters.kt
@@ -83,7 +83,7 @@ fun floatWithVisibilities(textView: TextView, value: Float?, prefix: String? = n
 }
 
 @BindingAdapter(value = ["price"])
-fun price(textView: TextView, price: Float?) {
+fun price(textView: TextView, price: Double?) {
     textView.text = if (
             price != null
             && price > 0F
